feat: qr 파일 병합
